[Fix] RankingDisplayer.java move package to root
