+++ src/main/java/jp/azisaba/lgw/rankingdisplayer/PluginConfig.java
@@ -8,7 +8,6 @@ import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.List;
 
-import jp.azisaba.lgw.rankingdisplayer.ranking.RankingDisplayer;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
